+++ include/openPMD/openPMD.hpp
@@ -49,6 +49,8 @@ namespace openPMD {}
 
 #include "openPMD/IO/AccessType.hpp"
 
+#include "openPMD/auxiliary/Date.hpp"
+#include "openPMD/auxiliary/Deprecated.hpp"
 #include "openPMD/auxiliary/OutOfRangeMsg.hpp"
 #include "openPMD/auxiliary/ShareRaw.hpp"
 #include "openPMD/auxiliary/Variant.hpp"